feat: remove old scheduled deliveries and flag
